Leaderboard

+++ app/model/AnswerContent.java
@@ -4,11 +4,11 @@ public class AnswerContent {
 	private int id;
 	private String time;
 	private String nickname;
-	private int votes;
+	private Float votes;
 	private String text;
 	
 	
-	public AnswerContent(String time, String text, String nickname, int votes, int id){
+	public AnswerContent(String time, String text, String nickname, Float votes, int id){
 		this.time = time;
 		this.nickname = nickname;
 		this.votes = votes;
